Refactor and Add postView Component
